feat: refactor stats into biller
