+++ pkg/biller/biller.go
@@ -9,10 +9,8 @@ import (
 
 	"github.com/docker/docker/api/types"
 	"github.com/docker/docker/client"
-	"github.com/hackclub/hack-as-a-service/pkg/api/apps"
 	"github.com/hackclub/hack-as-a-service/pkg/db"
 	"github.com/hackclub/hack-as-a-service/pkg/dokku"
-	"github.com/shopspring/decimal"
 	"gorm.io/gorm"
 )
 
@@ -81,41 +79,19 @@ func biller(app db.App, team db.Team, stream types.ContainerStats) {
 	for lines.Scan() {
 		line := lines.Text()
 		// log.Printf("Got line: %s\n", line)
-		var stat apps.Stats
+		var stat Stats
 		if err := json.Unmarshal([]byte(line), &stat); err != nil {
 			log.Printf("Error decoding json: %+v\n", err)
 			break
 		}
-		// log.Printf("Stat = %+v\n", stat)
-		// https://docs.docker.com/engine/api/v1.41/#operation/ContainerStats
-		// mem_usage := decimal.NewFromInt(stat.MemoryStats.Usage).Sub(decimal.NewFromInt(stat.MemoryStats.Stats.Cache)).Div(decimal.NewFromInt(stat.MemoryStats.Limit))
-		cpu_usage := decimal.NewFromInt(stat.CpuStats.CpuUsage.TotalUsage).Sub(decimal.NewFromInt(stat.PrecpuStats.CpuUsage.TotalUsage))
-		x := decimal.NewFromInt(stat.CpuStats.SystemCpuUsage).Sub(decimal.NewFromInt(stat.PrecpuStats.SystemCpuUsage))
-		if !x.IsZero() {
-			cpu_usage = cpu_usage.Div(x)
-		}
-		cpu_usage = cpu_usage.Mul(decimal.NewFromInt(int64(len(stat.CpuStats.CpuUsage.PercpuUsage))))
 
-		// Pricing (https://hackclub.slack.com/archives/C01N3B30TFB/p1619626597140600)
-		cpu_cost := decimal.RequireFromString("0.000001929012345679").Mul(cpu_usage)
-		cpu_cost = decimal.Max(decimal.Zero, cpu_cost)
-		// cpu_cost := decimal.RequireFromString("0.5")
-
-		// log.Printf("[%s][hn]%+v", app.ShortName, cpu_cost)
+		output := stat.Process()
+		expense := output.price()
 
 		// Accrue this new cost to the team
-		result := db.DB.Model(&team).Update("expenses", gorm.Expr("expenses + ?::decimal", cpu_cost))
+		result := db.DB.Model(&team).Update("expenses", gorm.Expr("expenses + ?::decimal", expense))
 		if result.Error != nil {
 			log.Printf("Error: %+v\n", result.Error)
 		}
-		// log.Printf("Result: %+v [team id = %d]\n", result, team.ID)
-
-		// output2 :=
-		// output := apps.WsOutput{
-		// 	Timestamp:       stat.Read,
-		// 	MemUsagePercent: mem_usage,
-		// 	CpuUsagePercent: cpu_usage,
-		// }
-		// log.Printf("Output = %+v\n", output)
 	}
 }
